requirements.txt
openai
python-decouple
feedparser
numpy
certifi
ssl
urllib3
google-generativeai